+++ src/components/survey/SurveySingleItemView/ResponseComponent/InputTypes/MultipleChoiceGroup.tsx
@@ -145,7 +145,9 @@ const MultipleChoiceGroup: React.FC<MultipleChoiceGroupProps> = (props) => {
       switch (option.role) {
         // TODO: handle composite option types, when contains different inputs
         case 'option':
-          labelComponent = <label htmlFor={optionKey}>
+          labelComponent = <label htmlFor={optionKey}
+            className="flex-grow-1 cursor-pointer"
+          >
             {renderFormattedContent(option, props.languageCode, 'cursor-pointer')}
           </label>
           break;